requirements.txt
tensorflow
numpy
matplotlib
seaborn
scikit-learn
Pillow
tabulate
gradio